requirements.txt
Flask==2.1.3
python-dotenv==1.0.1
Requests==2.32.3
web3==5.29.0
